az204/00_compute/azappservices/simple-storage-service/app/requirements.txt
python-dotenv

azure-storage-blob
Flask==2.0.2
gunicorn
